# Conflicts:
#	src/include/pop3.h
#	src/util/pop3_commands.c
#	src/util/pop3_files.c

+++ src/parsers/parser.c
@@ -3,10 +3,8 @@
 #include <stdbool.h>
 #include <stdlib.h>
 #include <string.h>
-#include <assert.h>
 
-#include <parser.h>
-#include <util.h>
+#include "parser.h"
 
 /* CDT del parser */
 struct parser {
@@ -63,7 +61,7 @@ parser_feed(struct parser *p, const uint8_t c) {
         const int when = state[i].when;
         if (state[i].when <= 0xFF) {
             matched = (c == when);
-        } else if(state[i].when == ANY) {
+        } else if((unsigned)state[i].when == ANY) {
             matched = true;
         } else if(state[i].when > 0xFF) {
             matched = (type & when);
